Moved amount type and associated classes to core bundle
